requirements.txt
flask==2.3.3
requests>=2.25.1
python-dotenv>=0.19.0
